[AIRFLOW-5777] Migrate AWS DynamoDB components to /providers/aws [AIP-21]
