+++ airflow/providers/aws/hooks/dynamodb.py
