+++ tests/providers/aws/hooks/test_dynamodb.py
@@ -21,7 +21,7 @@
 import unittest
 import uuid
 
-from airflow.contrib.hooks.aws_dynamodb_hook import AwsDynamoDBHook
+from airflow.providers.aws.hooks.dynamodb import AwsDynamoDBHook
 
 try:
     from moto import mock_dynamodb2
